HomeWork6/requirements.txt
fastapi
uvicorn
openai-whisper
openai
transformers
torch
gTTS
sympy
arxiv
python-multipart
langgraph
langchain-openai
langchain-core